refactor: Move code from list_view::component.rs to list_view.rs.
