+++ src/rust/ensogl/lib/components/src/drop_down_menu.rs
@@ -248,7 +248,7 @@ impl DropDownMenu {
                 // Align the top of the menu to the bottom of the icon.
                 model.selection_menu.set_position_y(-menu_size.y/2.0-icon_size.y/2.0-menu_offset_y);
                 // Align the right of the menu to the right of the icon.
-                let offfset_y = -menu_size.x/2.0+icon_size.x/2.0-list_view::component::SHADOW_PX/2.0;
+                let offfset_y = -menu_size.x/2.0+icon_size.x/2.0-list_view::SHADOW_PX/2.0;
                 model.selection_menu.set_position_x(offfset_y);
             });
