+++ src/rust/ide/view/src/searcher.rs
@@ -202,7 +202,7 @@ impl View {
 
             eval height.value ((h)  model.set_height(*h));
             eval frp.show     ((()) height.set_target_value(SEARCHER_HEIGHT));
-            eval frp.hide     ((()) height.set_target_value(-list_view::component::SHADOW_PX));
+            eval frp.hide     ((()) height.set_target_value(-list_view::SHADOW_PX));
 
             is_selected         <- model.list.selected_entry.map(|e| e.is_some());
             displayed_doc       <- model.list.selected_entry.map(f!((id) model.documentation_for_entry(*id)));
